Update CP437.hs
One is supposed to accumulate a builder with a right fold. https://www.reddit.com/r/haskell/comments/35udfy/fixing_an_aeson_performance_bug_sometimes_the_old/cr94pdj 
I skipped that with this attempt, which unfolds the text directly. It seems 10 or more times as fast if I send /usr/share/dict/words through it. `unfoldr` is a pretty strong function with text since it is basically just replicates to the stream representation. But maybe there is something better.